trying to fix the issues with the concurrency tests

+++ libraries/kiwi/kiwi-triplestore/src/test/java/org/apache/marmotta/kiwi/test/MySQLConcurrencyTest.java
@@ -98,7 +98,7 @@ public class MySQLConcurrencyTest extends ConcurrencyTestBase {
     @AfterClass
     public static void dropDatabase() throws RepositoryException, SQLException, SailException {
         assertTrue(store.checkConsistency());
-    	if (store != null) {
+    	if (store != null && store.isInitialized()) {
 	        store.closeValueFactory(); // release all connections before dropping the database
 	        store.getPersistence().dropDatabase();
 	        repository.shutDown();
